Add tailwind as a global dep. Add default rollup and tailwind config. Remove tailwind deps on style-guide.
