+++ packages/style-guide/package.json
@@ -15,8 +15,6 @@
     "@rollup/plugin-node-resolve": "^11.0.0",
     "@rollup/plugin-typescript": "^8.0.0",
     "@tsconfig/svelte": "^1.0.0",
-    "autoprefixer": "^10.3.1",
-    "postcss": "^8.3.6",
     "rollup": "^2.3.4",
     "rollup-plugin-css-only": "^3.1.0",
     "rollup-plugin-livereload": "^2.0.0",
@@ -26,7 +24,6 @@
     "svelte": "^3.0.0",
     "svelte-check": "^1.0.0",
     "svelte-preprocess": "^4.0.0",
-    "tailwindcss": "^2.2.7",
     "tslib": "^2.0.0",
     "typescript": "^4.0.0"
   },
